- Jorrit fixed several warnings in CEL caused by a few places 	  where size_t was not fixed. 	- Jorrit added a new iEntityLoader interface which is implemented 	  by the celentity addon. Using that interface one can now 	  more easilly access the celentity addon to load entities from 	  XML description files.

+++ libs/celtool/stdpcimp.cpp
@@ -62,7 +62,7 @@ void celPcCommon::FirePropertyChangeCallback (int propertyId)
 
 bool celPcCommon::AddPropertyChangeCallback (iCelPropertyChangeCallback* cb)
 {
-  int idx = callbacks.Find (cb);
+  size_t idx = callbacks.Find (cb);
   if (idx != csArrayItemNotFound) return false;
   callbacks.Push (cb);
   return true;
